Added Script to Change Scenes
I used all the YokArt given to me to create a start screen. When you hit
start, it switches scenes. To do this, I added both scenes to the master
scene list in BuildSettings. Then I created a control script with a
simple function that I attached to the Canvas. The button then has a
onClick function that calls the Canvas functions, which my custom
function is a member of. There you just change the public string field
to the target scene and viola.